feat: add observability logging, metrics aggregation, and PostgreSQL compatibility fixes
- Add request timing middleware logging every API call with method, path, status, elapsed ms
- Add in-process MetricsService aggregating per-endpoint and per-query stats (avg, p95, call count)
- Expose GET /api/metrics and POST /api/metrics/reset endpoints
- Add slow query logging (configurable via LOG_SLOW_QUERY_MS env var, default 200ms)
- Instrument both PostgreSQL and SQLite drivers with query timing via performance.now()
- Fix PostgreSQL boolean type mismatch: normalize 0/1 literals and parameters to true/false
- Fix PostgreSQL BIGINT overflow: sort_timestamp, taken_at, file_size, total_size → BIGINT
- Fix PostgreSQL IS NOT ? syntax: use IS DISTINCT FROM for null-safe param comparison
- serializeAnimatedFlag returns boolean for postgres, integer for sqlite

+++ server/db/migrations/postgres/000001_baseline.sql
@@ -49,7 +49,7 @@ CREATE TABLE IF NOT EXISTS images (
   extension TEXT NOT NULL,
   relative_path TEXT NOT NULL UNIQUE,
   absolute_path TEXT NOT NULL,
-  file_size INTEGER NOT NULL,
+  file_size BIGINT NOT NULL,
   width INTEGER NOT NULL,
   height INTEGER NOT NULL,
   display_orientation INTEGER NULL,
@@ -60,8 +60,8 @@ CREATE TABLE IF NOT EXISTS images (
   checksum_or_fingerprint TEXT NOT NULL,
   mtime_ms DOUBLE PRECISION NOT NULL,
   first_seen_at TEXT NOT NULL,
-  sort_timestamp INTEGER NOT NULL,
-  taken_at INTEGER NULL,
+  sort_timestamp BIGINT NOT NULL,
+  taken_at BIGINT NULL,
   taken_at_source TEXT NULL,
   exif_json TEXT NULL,
   thumbnail_path TEXT NOT NULL,
@@ -103,7 +103,7 @@ CREATE TABLE IF NOT EXISTS folder_scan_state (
   signature TEXT NOT NULL,
   file_count INTEGER NOT NULL,
   max_mtime_ms DOUBLE PRECISION NOT NULL,
-  total_size INTEGER NOT NULL,
+  total_size BIGINT NOT NULL,
   updated_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
